DBconnectionPane is now working normally, and bug proof.
This being said, it should be possible to handle it differently (better)
because currently using the equivalent of an ok_cancel option pane.
(checking the fields before quitting made it more complicated)

The whole management of DBconnection should be rewritten with a list of
parameters shared between DBconnection and DBconnectionPane. => should
be superclasses of connections such as MySQLDBconnection or
SQLiteDBconnection