added pentaho to the project.
Pentaho was added to the project. the software is in the git ignore for
now though.

on top of this: the item instances are now updated when the
transformation occurs. This being said, the saving file is not modified
yet. I am wondering if I should handle the savings differently => one
file per item => give the item the possibility to be saved even out of
the dataList.