the run button is now working.
I could run a transformation. I'm unhappy with the way I store the
transformations because it depends on the custom name of the item. I
want to control this name so I am thinking about using a comboBox
(adapted to each client, obviously)

I noticed that I need to manage the date attribute differently because
the date of a field can be a irrelevant compared to the date of the
transformation. (The ideal would be a field containing the date at which
the data was inserted). 
->My idea to deal with it for now is to use pentaho in order to find the
max Date and store it with pentaho also. Not sure if I can do this in
the same transformation as csv creation because of the parallel aspect
of it.


new features:
-The items are now able to generate the folders they need. This is done
at initialization and when new items are added

-The lastTransformationDate is now never null. instead it has a default
date that is something during year 1969

-The item now can return their jobFile, path, and same goes with saving.
The hierarchy for jobFile is
transformation/jobs/<itemType>/<itemName>/ktr-kjb
The method's hierarchy to do it was a little bit improved (see item and
abstract item to understand)

-the commandExecutor's main method is now useless. not removed but not
static anymore.
Aside of that, there is now a static constant, representing the path to
pentaho.

-the commandExecutor's buildCommand method was corrected. their was a
few problem with it (1 useless arg and 1 missing arg).

-the commandExecutor's execute(item) method now throws IOException when
there is no job file. (not sure if that's the good type of exception but
whatever...)

-MyItemView now saves the item when it is used (even if it fails)

-some modifications were done to the job. I adapted the paths so it
would work inside of this project. I added some steps to find the max
value, but no idea if it works