bug fixes
-The name of the files are now handled in a better way (in pentaho) so
that the date correctly appears in it, without creating useless folders.

-the generateFolders method in abstractItem was creating a useless node
in the hierarchy of savings. it doesn't anymore

-the time in abstractItem is now based on 24hours instead of 12 (hh->HH)

-myItemView now deletes all the existing job and create new ones when
the run button is pushed. for now, the new jobs are taken into the
sources folder (transformation/jobs/source). This helps developping new
features in pentaho and try them in eclipse. (just edit the
transformations in source folder)



It sounds like the data isn't correctly filtered whith the dates. (not
so sure about this.)