Make items unique.
the idea of this commit was to make items unique. To do so, we use a
unique identifier from the file identifiers/identifiers.txt and we give
the user the possibility to chose the identifier associated to the item
of data. For now, it is not really unique because the list of remaining
identifiers doesn't get updated when an item is created or removed (when
1 of the identifiers is used or freed)

another idea is to change the command line arguments to call a pentaho
job. we simply need to give kettle the name of the identifier (that will
be sent to anthony btw) and through that identifier, it will be able to
search the saving text file of the item to fetch the arguments.

-> Changes:

-New class: Identifiers, a singleton that manages a unique list of
remaining identifiers. many changes still need to be done, in order to
manage the app reboot and manual suppression of folders. (I shouldn't
bother too much about it. manual suppression is not always repairable)

A special pane is opened to chose the identifier when the user wants to
add an item to one of the lists. -> in the Item constructor -> now
throws cancel exception.

-New interface MySupplier: The supplier is now modified to throw
exceptions with the method get.

-AbstractDataList can't use the supplier anymore because the simple
constructor now forces to chose an Identifier. -> now makes call on
static method from concrete items. (not very proud of it...)

-AbstractDBconnection modified to take into account the identifier. this
being said, it's not yet able to create the commandLineArguments.

-AbstractItem modified to take into account these Identifiers instead of
the connectionName.

-FileList duplicates the information with itemSavingFolder. It should be
modified.

-Item interface: name() method deleted 

-MyItemView modified in order to take into account the Identifier
instead of name

-MySQLconnection modified to ignore the previous connectionName