identifiers successfully handled
The Identifiers class now has static methods free(Item) and use(Item)
which purpose is to allocate an identifier or free it to a specific item
(when this item is created or removed). the methods modified are init,
add and remove.




the abstract DataList was modified to manage the methods Identifiers.use
and Identifiers.free.  Now the identifiers are correctly assigned to
items.