code prototype for https communication + new Exception + bugfix
The pseudo code to connect to the API and to retrieve the job files was
written. This being said, saving the password is lacking. the next
commit should take care of this.

A small bug was resolved in the ApiConnectionPane that prevented the
displaying of the window.

An exception that is useless for now was created. The idea is to handle
the situation where the job files are not correctly retrieve.