managing of savings file (api connection ids)
the apiConnectionPane now closes if the connection is a succes. This
class is also more adapted to the modifications made to HttpManager


httpManager: saving the ids is now handled.
Also, the class now uses a boolean attribute that says if the connection
ids are initialized or not

This way, we will be able to use the default constructor without asking
the user to enter the ids a second time. (pretty necessary for an
automatic-periodical application)